Upgrade library to Angular9
