--- CHANGELOG.md
@@ -1,15 +1,0 @@
-CHANGELOG
-=========
-
-## HEAD (Unreleased)
-_(none)_
-
---------------------
-
-## 1.1.1 (2018-04-23)
-* Fix missing date error
-
-## 1.1.0 (2018-04-21)
-* Feature: Component can be disabled
-* Feature: Disable quartz
-
