--- docs/index.html
@@ -1,1 +1,0 @@
-<!doctype html><html lang="en"><head><meta charset="utf-8"><title>ngx-cron</title><meta name="viewport" content="width=device-width,initial-scale=1"><link rel="icon" type="image/x-icon" href="favicon.ico"><link type="text/css" rel="stylesheet" href="//unpkg.com/@swimlane/ngx-ui@20.2.0/release/index.css"><link href="styles.0532a5564212c0b098e5.bundle.css" rel="stylesheet"/></head><body><h1><a href="https://github.com/swimlane/ngx-cron">@swimlane/ngx-cron</a></h1><h2>🕒 User-friendly cron input</h2><div class="social-toolbox"><a class="github-button" href="https://github.com/swimlane/ngx-cron" data-size="large" data-count-href="/swimlane/ngx-cron/stargazers" data-show-count="true" data-count-aria-label="# stargazers on GitHub" aria-label="Star swimlane/ngx-dnd on GitHub">Star</a><div class="addthis_inline_share_toolbox"></div></div><app-root>Loading...</app-root><script async defer="defer" src="https://s7.addthis.com/js/300/addthis_widget.js#pubid=ra-583eec22a9881701"></script><script async defer="defer" src="https://buttons.github.io/buttons.js"></script><script type="text/javascript" src="inline.318b50c57b4eba3d437b.bundle.js"></script><script type="text/javascript" src="polyfills.015c48ae49c7ee0f836a.bundle.js"></script><script type="text/javascript" src="main.12dd6f6052b317ad5f0e.bundle.js"></script></body></html>
