--- docs/styles.0532a5564212c0b098e5.bundle.css
@@ -1,1 +1,0 @@
-body{background:#1c2029;color:#cfcfcf;font-family:Lato;padding:10px;margin:10px}[hidden]{display:none!important}.social-toolbox{position:absolute;top:20px;right:30px;z-index:999;display:-webkit-box;display:-ms-flexbox;display:flex;-webkit-box-align:center;-ms-flex-align:center;align-items:center;-webkit-box-pack:end;-ms-flex-pack:end;justify-content:flex-end;width:75%;width:calc(75% - 10px)}.social-toolbox iframe{margin-right:5px}.social-toolbox .addthis_inline_share_toolbox{margin-top:5px}body>h1,body>h2{margin:0 0 0 10px}
