+++ package.json
@@ -1,91 +1,56 @@
 {
-  "name": "@swimlane/ngx-cron",
-  "version": "1.1.1",
-  "license": "MIT",
+  "name": "ngx-cron",
+  "version": "0.0.0",
   "scripts": {
     "ng": "ng",
     "start": "ng serve",
-    "build": "npm-run-all build:*",
-    "build:demo": "ng build --prod -op docs",
-    "build:package": "ng-packagr -p ng-package.json",
+    "build": "ng build",
     "test": "ng test",
-    "test:single": "ng test --single-run",
-    "test:ci": "ng test --single-run --no-progress --browser=ChromeNoSandbox",
     "lint": "ng lint",
-    "e2e": "ng e2e",
-    "changelog": "chg release -y && git add -A CHANGELOG.md",
-    "preversion": "npm-run-all lint test:single",
-    "version": "npm-run-all build changelog"
+    "e2e": "ng e2e"
   },
-  "peerDependencies": {
-    "@angular/core": "^5.2.0",
-    "@angular/forms": "^5.2.0",
-    "@angular/http": "^5.2.0",
-    "@angular/platform-browser": "^5.2.0",
-    "@angular/platform-browser-dynamic": "^5.2.0",
-    "@angular/router": "^5.2.0",
-    "core-js": "^2.5.5",
-    "rxjs": "^5.5.10",
-    "zone.js": "^0.8.19",
-    "cronstrue": "^1.2.0",
-    "@swimlane/ngx-ui": "^20.2.0",
-    "@angular/animations": "^5.2.0",
-    "@angular/common": "^5.2.0",
-    "@angular/compiler": "^5.2.0"
+  "private": true,
+  "dependencies": {
+    "@angular/animations": "~9.1.4",
+    "@angular/cdk": "^9.2.1",
+    "@angular/common": "~9.1.4",
+    "@angular/compiler": "~9.1.4",
+    "@angular/core": "~9.1.4",
+    "@angular/forms": "~9.1.4",
+    "@angular/platform-browser": "~9.1.4",
+    "@angular/platform-browser-dynamic": "~9.1.4",
+    "@angular/router": "~9.1.4",
+    "@swimlane/ngx-ui": "^29.1.0",
+    "codemirror": "^5.49.2",
+    "cronstrue": "^1.94.0",
+    "moment-timezone": "^0.5.31",
+    "ngx-moment": "^5.0.0",
+    "resize-observer-polyfill": "^1.5.1",
+    "rxjs": "~6.5.4",
+    "tslib": "^1.10.0",
+    "zone.js": "~0.10.2"
   },
   "devDependencies": {
-    "@angular/animations": "^5.2.0",
-    "@angular/cli": "~1.7.4",
-    "@angular/common": "^5.2.0",
-    "@angular/compiler": "^5.2.0",
-    "@angular/compiler-cli": "^5.2.0",
-    "@angular/core": "^5.2.0",
-    "@angular/forms": "^5.2.0",
-    "@angular/http": "^5.2.0",
-    "@angular/language-service": "^5.2.0",
-    "@angular/platform-browser": "^5.2.0",
-    "@angular/platform-browser-dynamic": "^5.2.0",
-    "@angular/router": "^5.2.0",
-    "@swimlane/ngx-ui": "^20.2.0",
-    "@types/jasmine": "~2.8.3",
-    "@types/jasminewd2": "~2.0.2",
-    "@types/node": "~6.0.60",
-    "chg": "^0.3.3",
-    "codelyzer": "^4.0.1",
-    "core-js": "^2.5.5",
-    "cronstrue": "^1.2.0",
-    "jasmine-core": "~2.8.0",
+    "@angular-devkit/build-angular": "~0.901.4",
+    "@angular-devkit/build-ng-packagr": "~0.901.10",
+    "@angular/cli": "~9.1.4",
+    "@angular/compiler-cli": "~9.1.4",
+    "@angular/language-service": "~9.1.4",
+    "@types/node": "^12.11.1",
+    "@types/jasmine": "~3.5.0",
+    "@types/jasminewd2": "~2.0.3",
+    "codelyzer": "^5.1.2",
+    "jasmine-core": "~3.5.0",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~2.0.0",
-    "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "^1.2.1",
-    "karma-jasmine": "~1.1.0",
-    "karma-jasmine-html-reporter": "^0.2.2",
-    "ng-packagr": "^2.4.2",
-    "npm-run-all": "^4.1.2",
-    "protractor": "~5.1.2",
-    "rxjs": "^5.5.10",
-    "ts-node": "~4.1.0",
-    "tslint": "~5.9.1",
-    "tslint-config-swimlane": "^3.0.4",
-    "typescript": "~2.5.3",
-    "zone.js": "^0.8.19"
-  },
-  "description": "User-friendly cron input",
-  "directories": {
-    "doc": "docs"
-  },
-  "dependencies": {},
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/swimlane/ngx-cron.git"
-  },
-  "keywords": [
-    "cron"
-  ],
-  "author": "J. Harshbarger",
-  "bugs": {
-    "url": "https://github.com/swimlane/ngx-cron/issues"
-  },
-  "homepage": "https://github.com/swimlane/ngx-cron#readme"
+    "karma": "~5.0.0",
+    "karma-chrome-launcher": "~3.1.0",
+    "karma-coverage-istanbul-reporter": "~2.1.0",
+    "karma-jasmine": "~3.0.1",
+    "karma-jasmine-html-reporter": "^1.4.2",
+    "ng-packagr": "^9.0.0",
+    "protractor": "~5.4.3",
+    "ts-node": "~8.3.0",
+    "tslint": "~6.1.0",
+    "typescript": "~3.8.3"
+  }
 }
