--- protractor-ci.conf.js
@@ -1,10 +1,0 @@
-const config = require('./protractor.conf').config;
-
-config.capabilities = {
-  browserName: 'chrome',
-  chromeOptions: {
-    args: ['--no-sandbox']
-  }
-};
-
-exports.config = config;
