--- public_api.ts
@@ -1,1 +1,0 @@
-export * from './src/app/modules/ngx-cron/ngx-cron.module';
