--- src/app/app.component.css
@@ -1,10 +1,0 @@
-section.component {
-  padding: 10px 20px;
-  margin: 10px 0px;
-  box-shadow: 0 7px 8px -4px rgba(0, 0, 0, 0.2), 0 12px 17px 2px rgba(0, 0, 0, 0.14), 0 5px 22px 4px rgba(0, 0, 0, 0.12);
-  border: 1px solid #313847;
-}
-
-ngx-section {
-  margin: 10px 0px;
-}
