--- src/app/modules/ngx-cron/ngx-cron.component.html
@@ -1,119 +1,0 @@
-<div>
-  <div *ngIf="_allowedPeriods.length > 1">
-    <ng-container *ngFor="let key of _allowedPeriods">
-      <ngx-button class="btn"
-        [class.btn-primary]="cronData.period === key"
-        (click)="cronData.period = key; cronDataChanged()"
-        [disabled]="disabled">
-        {{key}}
-      </ngx-button>
-    </ng-container>
-  </div>
-
-  <div class="selections">
-    <span *ngIf="cronData.period === 'Custom'">
-      <ngx-input
-        [(ngModel)]="cron"
-        [hint]="cronData.description"
-        [disabled]="disabled || disableCustomInput">
-      </ngx-input>
-    </span>
-
-    <span *ngIf="cronData.period === 'Hourly'">
-      At
-      <ngx-input
-        type="number"
-        [(ngModel)]="cronData.minute"
-        (change)="cronDataChanged()"
-        min="0"
-        max="59"
-        [disabled]="disabled">
-      </ngx-input>
-      <span [ngPlural]="cronData.minute">
-        <ng-template ngPluralCase="=1">minute</ng-template>
-        <ng-template ngPluralCase="other">minutes</ng-template>
-      </span> past the hour
-    </span>
-
-    <span class="every" *ngIf="cronData.period === 'Secondly'">
-      Every 
-      <ngx-input
-        type="number"
-        [ngModel]="cronData.secondInterval"
-        (ngModelChange)="cronData.secondInterval = +$event"
-        (change)="cronDataChanged()"
-        min="1"
-        max="9999"
-        [disabled]="disabled">
-      </ngx-input>
-      <span [ngPlural]="cronData.secondInterval">
-        <ng-template ngPluralCase="=1">second</ng-template>
-        <ng-template ngPluralCase="other">seconds</ng-template>
-      </span>
-    </span>
-
-    <span class="every" *ngIf="cronData.period === 'Minutely'">
-      Every 
-      <ngx-input
-        type="number"
-        [ngModel]="cronData.minuteInterval"
-        (ngModelChange)="cronData.minuteInterval = +$event"
-        (change)="cronDataChanged()"
-        min="1"
-        max="9999"
-        [disabled]="disabled">
-      </ngx-input>
-      <span [ngPlural]="cronData.minuteInterval">
-        <ng-template ngPluralCase="=1">minute</ng-template>
-        <ng-template ngPluralCase="other">minutes</ng-template>
-      </span>
-    </span>
-
-    <span *ngIf="cronData.period === 'Daily' || cronData.period === 'Weekly' || cronData.period === 'Monthly' || cronData.period === 'Yearly'">
-      At
-      <ngx-date-time
-        [(ngModel)]="cronData.time"
-        (change)="cronDataChanged()"
-        inputType="time"
-        [disabled]="disabled">
-      </ngx-date-time>
-    </span>
-
-    <span *ngIf="cronData.period === 'Monthly' || cronData.period === 'Yearly'">
-      , on day 
-      <ngx-input
-        type="number"
-        [(ngModel)]="cronData.day"
-        (change)="cronDataChanged()"
-        min="1"
-        [max]="cronData.period === 'Yearly' ? cronData.daysMax : 31"
-        [disabled]="disabled">
-      </ngx-input>
-      of the month
-    </span>
-
-    <span *ngIf="cronData.period === 'Yearly'">
-      , only in
-      <ngx-select
-        [filterable]="false"
-        [ngModel]="[cronData.month]"
-        [disabled]="disabled"
-        (change)="cronData.month = $event[0]; cronDataChanged()"
-        [allowClear]="false">
-        <ngx-select-option *ngFor="let month of months" [name]="month" [value]="month"></ngx-select-option>
-      </ngx-select>
-    </span>
-
-    <span *ngIf="cronData.period === 'Weekly'">
-      , only on
-      <ngx-select
-        [filterable]="false"
-        [ngModel]="[cronData.weekday]"
-        (change)="cronData.weekday = $event[0]; cronDataChanged()"
-        [allowClear]="false"
-        [disabled]="disabled">
-        <ngx-select-option *ngFor="let day of dows" [name]="day" [value]="day"></ngx-select-option>
-      </ngx-select>
-    </span>
-  </div>
-</div>
