--- src/app/modules/ngx-cron/ngx-cron.component.scss
@@ -1,58 +1,0 @@
-
-ngx-cron-input {
-  display: block;
-  border: 1px solid transparent;
-  margin-top: 30px;
-  padding: 10px;
-
-  .ngx-input, .ngx-select, .ngx-select-wrap, .ngx-select-flex-wrap, .ngx-select-input {
-    display: inline-block;
-    width: auto;
-    min-width: initial;
-    margin: 0;
-    padding: 0;
-  }
-
-  .ngx-input {
-    padding-top: 5px;
-  }
-
-  .ngx-select-dropdown {
-    min-width: 300px;
-  }
-
-  ngx-date-time {
-    display: inline-block;
-    width: auto;
-    min-width: initial;
-    vertical-align: middle;
-
-    .ngx-input {
-      padding-top: 5px;
-      margin-top: 0;
-
-      .ngx-input-wrap .ngx-input-hint {
-        height: 0;
-        min-height: 0;
-      }
-    }
-
-    .calendar-dialog-btn {
-      font-size: 0.8em;
-      margin-top: -8px;
-    }
-  }
-
-  ngx-button.selected {
-    background-color: red;
-  }
-
-  .period-selection, .selections {
-    margin-top: 16px;
-    height: 70px;
-  }
-
-  &.invalid {
-    border: 1px solid red;
-  }
-}
