--- src/styles.css
@@ -1,39 +1,0 @@
-/* You can add global styles to this file, and also import other style files */
-
-body {
-  background: #1c2029;
-  color: #cfcfcf;
-  font-family: Lato;
-  padding: 10px;
-  margin: 10px;
-}
-
-[hidden] {
-  display: none !important;
-}
-.social-toolbox {
-  position: absolute;
-  top: 20px;
-  right: 30px;
-  z-index: 999;
-  display: flex;
-  align-items: center;
-  justify-content: flex-end;
-  width: 75%;
-  width: calc(75% - 10px);
-}
-.social-toolbox iframe {
-  margin-right: 5px;
-}
-.social-toolbox .addthis_inline_share_toolbox {
-  margin-top: 5px;
-}
-
-body > h1 {
-  margin-top: 0;
-  margin-bottom: 20px;
-}
-
-body > h1, body > h2 {
-  margin: 0 0 0 10px;
-}
