--- todo.md
@@ -1,7 +1,0 @@
-# Todo list
-
-_\( managed using [todo-md](https://github.com/Hypercubed/todo-md) \)_
-
-- [ ] i18n support
-- [ ] tests
-- [ ] cron pipe?
